add final version

+++ main.py
@@ -1,13 +1,39 @@
 from tkinter import *
+from tkinter import messagebox
+from random import randint, choice, shuffle
+import pyperclip
 
 
-def add_to_file():
+def generate_password():
+    letters = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z']
+    numbers = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']
+    symbols = ['!', '#', '$', '%', '&', '(', ')', '*', '+']
+
+    password_letters = [choice(letters) for _ in range(randint(8, 10))]
+    password_symbols = [choice(symbols) for _ in range(randint(2, 4))]
+    password_numbers = [choice(numbers) for _ in range(randint(2, 4))]
+
+    password_list = password_letters + password_symbols + password_numbers
+    shuffle(password_list)
+    password = "".join(password_list)
+    password_entry.insert(0,password)
+    pyperclip.copy(password)
+
+
+def save():
+
     website = website_entry_field.get()
     username = username_password_entry.get()
     password = password_entry.get()
-    with open("data.txt", mode="a") as data_file:
-        data_file.write(f"{website} | {username} | {password}\n")
-        delete_data()
+
+    if len(website) == 0 or len(password) == 0:
+        messagebox.showinfo(title="Ooops", message="Please don't leave any fields empty!")
+    else:
+        is_ok = messagebox.askokcancel(title=website, message=f"These are the details entered:\nEmail: {username} \nPassword: {password} \nIs it ok to save?")
+        if is_ok:
+            with open("data.txt", mode="a") as data_file:
+                data_file.write(f"{website} | {username} | {password}\n")
+                delete_data()
 
 
 def delete_data():
@@ -32,8 +58,8 @@ website_entry_field = Entry(width=35)
 username_password_entry = Entry(width=21)
 password_entry = Entry(width=35)
 
-generate_button = Button(text="Generate Password")
-add_button = Button(text="Add", width=35, command=add_to_file)
+generate_button = Button(text="Generate Password", command=generate_password)
+add_button = Button(text="Add", width=35, command=save)
 
 website_label.grid(column=0, row=3)
 username_email_label.grid(column=0, row=4)
